+++ fe/src/components/analysis/Budget/BudgetView.tsx
@@ -3,8 +3,8 @@ import { Calculator, Edit, ListPlus } from 'lucide-react';
 import AnalysisEmpty from '@/components/analysis/common/AnalysisEmpty';
 import AnalysisLoading from '@/components/analysis/common/AnalysisLoading';
 import AnalysisSection from '@/components/analysis/common/AnalysisSection';
-import BudgetRecommendDialog from './BudgetRecommendDialog';
-import BudgetSetupDialog from '@/components/analysis/BudgetSetupDialog';
+import BudgetRecommendDialog from '@/components/analysis/Budget/BudgetRecommendDialog';
+import BudgetSetupDialog from '@/components/analysis/Budget/BudgetSetupDialog';
 import { Button } from '@/components/ui/button';
 import CategoryBudgetSetting from '@/components/analysis/CategoryBudgetSetting';
 import ConfirmDialog from '@/components/analysis/common/ConfirmDialog';
